SDK-2425.  Fix scheduled meetings compilation issues (fix pending issues)

See merge request sdk/sdk!3935

+++ src/megaapi_impl.cpp
@@ -33749,7 +33749,7 @@ MegaScheduledMeetingListPrivate::~MegaScheduledMeetingListPrivate()
 
 unsigned long MegaScheduledMeetingListPrivate::size() const
 {
-    return mList.size();
+    return static_cast<unsigned long>(mList.size());
 }
 
 MegaScheduledMeetingListPrivate* MegaScheduledMeetingListPrivate::copy() const
